feat: db migration

+++ src/utils.rs
@@ -1,3 +1,3 @@
 pub fn mask_user(user: i64) -> String {
-    format!("{:x}", md5::compute(user.to_le_bytes()))
+    format!("{:x}", md5::compute((user as u128).to_le_bytes()))
 }
